[update]post_hashtagsモデル

+++ db/migrate/20211115034304_create_post_hashtags.rb
@@ -1,7 +1,7 @@
 class CreatePostHashtags < ActiveRecord::Migration[5.2]
   def change
     create_table :post_hashtags do |t|
-      t.integer :post_id
+      t.bigint :post_id
       t.integer :hashtag_id
 
       t.timestamps
